Move sentences controller into person directory
And fix some formatting.
